+++ src/potent/plan.py
@@ -8,6 +8,7 @@ from rich.console import Console
 from rich.tree import Tree
 
 from potent.operations._base import AbsDirPath, Status
+from potent.operations.change_pr_status import ChangePrStatus
 from potent.operations.create_pr import CreatePR
 from potent.operations.enable_automerge import EnableAutomerge
 from potent.operations.git_add import GitAdd
@@ -172,6 +173,7 @@ class Plan(BaseModel):
                 EnableAutomerge,
                 RawCommand,
                 ManualConfirmation,
+                ChangePrStatus,
                 # OPERATIONS ^
             ],
             Field(discriminator="slug"),